refactor: move utilities to specific files
