+++ typescript/src/arguments/song.ts
@@ -1,8 +1,8 @@
 import { GuildSettings, readSettings } from '#lib/database';
 import { LanguageKeys } from '#lib/i18n/languageKeys';
 import type { GuildMessage } from '#lib/types';
+import { count, filter, map, take } from '#utils/common';
 import { isDJ } from '#utils/functions';
-import { count, filter, map, take } from '#utils/iterator';
 import { fetch, FetchResultTypes } from '@sapphire/fetch';
 import { Args, Argument, ArgumentContext } from '@sapphire/framework';
 import { parseURL } from '@sapphire/utilities';
