+++ typescript/src/audio/commands/Music/exportqueue.ts
@@ -1,7 +1,7 @@
 import { MusicCommand, QueueEntry, requireQueueNotEmpty } from '#lib/audio';
 import { LanguageKeys } from '#lib/i18n/languageKeys';
 import type { GuildMessage } from '#lib/types/Discord';
-import { map, prependIfNotNull, take } from '#utils/iterator';
+import { map, prependIfNotNull, take } from '#utils/common';
 import { ApplyOptions } from '@sapphire/decorators';
 import { serialize } from 'binarytf';
