+++ typescript/src/audio/lib/queue/Queue.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/unified-signatures */
 import { Events } from '#lib/types/Enums';
-import { map, reverse } from '#utils/iterator';
+import { map, reverse } from '#utils/common';
 import { Store } from '@sapphire/framework';
 import { Time } from '@sapphire/time-utilities';
 import { isNullish } from '@sapphire/utilities';
