+++ typescript/src/commands/Admin/conf.ts
@@ -3,8 +3,8 @@ import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { SettingsMenu, SkyraCommand } from '#lib/structures';
 import type { GuildMessage } from '#lib/types';
 import { PermissionLevels } from '#lib/types/Enums';
+import { filter, map } from '#utils/common';
 import { requiresPermissions } from '#utils/decorators';
-import { filter, map } from '#utils/iterator';
 import { ApplyOptions } from '@sapphire/decorators';
 import { toTitleCase } from '@sapphire/utilities';
