+++ typescript/src/commands/Animal/kitty.ts
@@ -1,6 +1,7 @@
 import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { SkyraCommand } from '#lib/structures';
-import { getImageUrl, wrap } from '#utils/util';
+import { safeWrapPromise } from '#utils/common';
+import { getImageUrl } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { fetch, FetchResultTypes } from '@sapphire/fetch';
 import { Message, MessageEmbed } from 'discord.js';
@@ -15,7 +16,7 @@ import { Message, MessageEmbed } from 'discord.js';
 })
 export class UserCommand extends SkyraCommand {
 	public async run(message: Message) {
-		const result = await wrap(fetch<AwsRandomCatResult>('https://aws.random.cat/meow', FetchResultTypes.JSON));
+		const result = await safeWrapPromise(fetch<AwsRandomCatResult>('https://aws.random.cat/meow', FetchResultTypes.JSON));
 		return message.send(
 			new MessageEmbed()
 				.setColor(await this.context.db.fetchColor(message))
