+++ typescript/src/commands/Announcement/announcement.ts
@@ -3,9 +3,9 @@ import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { SkyraCommand } from '#lib/structures';
 import type { GuildMessage } from '#lib/types';
 import { Events, PermissionLevels } from '#lib/types/Enums';
+import { map } from '#utils/common';
 import { BrandingColors } from '#utils/constants';
 import { promptConfirmation } from '#utils/functions';
-import { map } from '#utils/iterator';
 import { announcementCheck, DetailedMentionExtractionResult, extractDetailedMentions } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { RESTJSONErrorCodes } from 'discord-api-types/v6';
