+++ typescript/src/events/guilds/guildUpdateNotify.ts
@@ -2,7 +2,7 @@ import { GuildSettings, readSettings, writeSettings } from '#lib/database';
 import { LanguageKeys } from '#lib/i18n/languageKeys';
 import { Colors } from '#lib/types/Constants';
 import { toChannelsArray } from '#utils/bits';
-import { differenceArray, differenceBitField } from '#utils/comparators';
+import { differenceArray, differenceBitField } from '#utils/common/comparators';
 import { ApplyOptions } from '@sapphire/decorators';
 import { Event, EventOptions, Events } from '@sapphire/framework';
 import { Time } from '@sapphire/time-utilities';
