+++ typescript/src/events/reactions/rawMessageReactionAdd.ts
@@ -1,6 +1,7 @@
 import { Events } from '#lib/types/Enums';
+import { isGuildBasedChannel } from '#utils/functions';
 import type { LLRCData } from '#utils/LongLivingReactionCollector';
-import { isTextBasedChannel, resolveEmoji } from '#utils/util';
+import { resolveEmoji } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { Event, EventOptions } from '@sapphire/framework';
 import { GatewayDispatchEvents, GatewayMessageReactionAddDispatch } from 'discord-api-types/v6';
@@ -10,7 +11,7 @@ import type { TextChannel } from 'discord.js';
 export class UserEvent extends Event {
 	public run(raw: GatewayMessageReactionAddDispatch['d']) {
 		const channel = this.context.client.channels.cache.get(raw.channel_id) as TextChannel | undefined;
-		if (!channel || !isTextBasedChannel(channel) || !channel.readable) return;
+		if (!channel || !isGuildBasedChannel(channel) || !channel.readable) return;
 
 		const data: LLRCData = {
 			channel,
