+++ typescript/src/lib/birthday/utils.ts
@@ -1,5 +1,5 @@
 import type { ScheduleEntity } from '#lib/database';
-import { filter, first } from '#utils/iterator';
+import { filter, first } from '#utils/common';
 import { Store } from '@sapphire/pieces';
 import { BirthdayScheduleEntity, DateWithOptionalYear, Month } from './types';
