+++ typescript/src/lib/util/common/index.ts
@@ -1,2 +1,4 @@
+export * from './comparators';
 export * from './guards';
+export * from './iterators';
 export * from './promises';
