+++ typescript/src/lib/util/common/promises.ts
@@ -1,4 +1,4 @@
-import { Store } from '@sapphire/framework';
+import { err, ok, Result, Store } from '@sapphire/framework';
 import { isThenable, Awaited } from '@sapphire/utilities';
 import type { RESTJSONErrorCodes } from 'discord-api-types/v6';
 import { DiscordAPIError } from 'discord.js';
@@ -16,4 +16,50 @@ export async function resolveOnErrorCodes<T>(promise: Promise<T>, ...codes: read
 
 export function floatPromise(promise: Awaited<unknown>) {
 	if (isThenable(promise)) promise.catch((error: Error) => container.logger.fatal(error));
+}
+
+/**
+ * Read a stream and resolve to a buffer.
+ * @param stream The readable stream to read
+ */
+export async function streamToBuffer(stream: NodeJS.ReadableStream): Promise<Buffer> {
+	const data: Buffer[] = [];
+	for await (const buffer of stream) data.push(buffer as Buffer);
+	return Buffer.concat(data);
+}
+
+export interface ReferredPromise<T> {
+	promise: Promise<T>;
+	resolve(value?: T): void;
+	reject(error?: Error): void;
+}
+
+/**
+ * Create a referred promise.
+ */
+export function createReferPromise<T>(): ReferredPromise<T> {
+	let resolve: (value: T) => void;
+	let reject: (error?: Error) => void;
+	const promise: Promise<T> = new Promise((res, rej) => {
+		resolve = res;
+		reject = rej;
+	});
+
+	// noinspection JSUnusedAssignment
+	return { promise, resolve: resolve!, reject: reject! };
+}
+
+/**
+ * Wraps the result of a promise into a Result<T, E>, resulting on Ok<T> when successful, and Err<E> when erroneous.
+ * This function's purpose is to create promises that do not reject, allowing code to be written without try/catch or
+ * {@link Promise.catch} callbacks.
+ * @param promise The promise to wrap into a Result<T, E>
+ * @returns A promise resolving to a Result.
+ */
+export async function safeWrapPromise<T, E = Error>(promise: Promise<T>): Promise<Result<T, E>> {
+	try {
+		return ok(await promise);
+	} catch (error) {
+		return err(error);
+	}
 }
